+++ RELEASE-NOTES.md
@@ -1,2 +1,3 @@
 ## 0.1:
 
+* #1 Initial client implementation